terminal parsing tests
The commit also fixes several flaws in the fake_lexer implementation,
used by the parser tests
